adjustment of some json models

+++ src/data_models/data_object.py
@@ -3,16 +3,16 @@ from pydantic import BaseModel, conint
 
 
 class DataObjectMeasure(BaseModel):
-    Value: float
-    Invalidity: Literal[0, 1, 2]  # Nell adu stato deve essere solo 1 o 2
-    ErrorCode: Literal[0, 1, 2, 3, 4, 5, 6, 7, 8]  # Nell'Adu stato deve essere nullo
+    Value: int
+    Invalidity: Literal[0, 1, 2]
+    ErrorCode: Literal[0, 1, 2, 3, 4, 5, 6, 7, 8]
     Timetag: conint(gt=0)  # type: ignore
 
 
-# TODO: inserire i check relativi a Invalidity ed Errorcode
+# TODO: inserire i check relativi a Invalidity ed Errorcode (p. 35 PAS)
 
 
 class DataObjectSateAlarm(BaseModel):
-    Value: float
+    Value: int
     Invalidity: bool
     Timetag: conint(gt=0)  # type: ignore
